+++ src/LevelSize.hpp
@@ -2,5 +2,5 @@
 
 class LevelSize {
 public:
-    static std::string getSizeString(size_t size);
+    static std::string getSizeString(uint64_t size);
 };
